[hwcomposer_backend] Build fixes for building against CM 10.2 headers

+++ hwcomposer/hwcomposer_backend.h
@@ -42,6 +42,8 @@
 #ifndef HWCOMPOSER_BACKEND_H
 #define HWCOMPOSER_BACKEND_H
 
+#include <sys/types.h>
+
 #include <hardware/hardware.h>
 #include <hardware/hwcomposer.h>
 
@@ -59,12 +61,12 @@
 // Evaluate "x", if it isn't NULL, print a warning
 #define HWC_PLUGIN_EXPECT_NULL(x) \
     { void *res; if ((res = (x)) != NULL) \
-        qWarning("QPA-HWC: %s in %s returned %x", (#x), __func__, res); }
+        qWarning("QPA-HWC: %s in %s returned %x", (#x), __func__, (unsigned int)res); }
 
 // Evaluate "x", if it is NULL, exit with a fatal error
 #define HWC_PLUGIN_ASSERT_NOT_NULL(x) \
     { void *res; if ((res = (x)) == NULL) \
-        qFatal("QPA-HWC: %s in %s returned %x", (#x), __func__, res); }
+        qFatal("QPA-HWC: %s in %s returned %x", (#x), __func__, (unsigned int)res); }
 
 // Evaluate "x", if it doesn't return zero, exit with a fatal error
 #define HWC_PLUGIN_ASSERT_ZERO(x) \
